requirements.txt
python-telegram-bot==20.5
requests==2.31.0
python-dotenv==1.0.0
requests
python-dotenv
anthropic
google-cloud-pubsub
google-cloud-storage
google-cloud-texttospeech


